add a testbench and fix the verilog
The verilog was almost correct. :)

I removed the "and" of the first logic cloud and did an or instead. It
makes for more useful data.

The testbench is very very simple. It feeds in 128 random bits. Since
the bits have logical operations performed on them, we'll need to check
that the correct bits are coming back out. Working on that next.

Signed-off-by: Stephano Cetola <stephanoc@gmail.com>