add wave.do and cleanup
Adding wave.do files so that the format of the waveforms is consistant.

REmove "front_end" folder. This folder is an exact copy of the "rtl"
folder, but with the addition of a wave.do and a run script. Adding a
symbolic link to the rtl folder instead of duplicating code. Note this
is a unix symbolic link (won't work in windows / mac).

Add run scripts.

Signed-off-by: Stephano Cetola <stephanoc@gmail.com>